+++ modules/task_1/sablin_a_gaussian_vert/gaussian_vert.h
@@ -11,6 +11,6 @@ struct rgb_coub {
 };
 std::vector<rgb_coub> getRandomImage(size_t rows, size_t columns);
 std::vector<rgb_coub> Gaussian_Filter_Seq(const std::vector<rgb_coub>& img,
-    size_t rows, size_t columns, const double sigma);
+    int rows, int columns, const double sigma);
 
 #endif  // MODULES_TASK_1_SABLIN_A_GAUSSIAN_VERT_GAUSSIAN_VERT_H_
